backend/requirements.txt
# Core Django
Django>=5.0,<5.1
djangorestframework>=3.14.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# API Documentation
drf-spectacular>=0.27.0

# Image handling
Pillow>=10.0.0

# CORS
django-cors-headers>=4.3.0

# Production server
gunicorn>=21.2.0

# Environment variables
python-dotenv>=1.0.0

# Testing
coverage>=7.3.0

# Utilities
django-extensions>=3.2.3
